pyproject.toml
[project]
name = "automation-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.75.0",
    "gitingest>=0.3.1",
    "gitpython>=3.1.46",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.15.0",
]
